add EventMove command, fix extra signal call in map border resize
