+++ include/editor.h
@@ -145,6 +145,8 @@ public:
 
     QUndoGroup editGroup; // Manages the undo history for each map
 
+    bool selectingEvent = false;
+
 private:
     void setConnectionItemsVisible(bool);
     void setBorderItemsVisible(bool, qreal = 1);
@@ -202,106 +204,6 @@ signals:
     void warpEventDoubleClicked(QString mapName, QString warpNum);
     void currentMetatilesSelectionChanged();
     void wheelZoom(int delta);
-};
-
-
-
-class DraggablePixmapItem : public QObject, public QGraphicsPixmapItem {
-    Q_OBJECT
-public:
-    DraggablePixmapItem(QPixmap pixmap): QGraphicsPixmapItem(pixmap) {
-    }
-    Editor *editor = nullptr;
-    Event *event = nullptr;
-    QGraphicsItemAnimation *pos_anim = nullptr;
-    DraggablePixmapItem(Event *event_, Editor *editor_) : QGraphicsPixmapItem(event_->pixmap) {
-        event = event_;
-        editor = editor_;
-        updatePosition();
-    }
-    bool active;
-    int last_x;
-    int last_y;
-    void updatePosition() {
-        int x = event->getPixelX();
-        int y = event->getPixelY();
-        setX(x);
-        setY(y);
-        if (editor->selected_events && editor->selected_events->contains(this)) {
-            setZValue(event->y() + 1);
-        } else {
-            setZValue(event->y());
-        }
-    }
-    void move(int x, int y);
-    void emitPositionChanged() {
-        emit xChanged(event->x());
-        emit yChanged(event->y());
-        emit elevationChanged(event->elevation());
-    }
-    void updatePixmap() {
-        QList<Event*> objects;
-        objects.append(event);
-        event->pixmap = QPixmap();
-        editor->project->loadEventPixmaps(objects);
-        this->updatePosition();
-        editor->redrawObject(this);
-        emit spriteChanged(event->pixmap);
-    }
-    void bind(QComboBox *combo, QString key) {
-        connect(combo, static_cast<void (QComboBox::*)(const QString &)>(&QComboBox::currentTextChanged),
-                this, [this, key](QString value){
-            this->event->put(key, value);
-        });
-        connect(this, &DraggablePixmapItem::onPropertyChanged,
-                this, [combo, key](QString key2, QString value){
-            if (key2 == key) {
-                combo->addItem(value);
-                combo->setCurrentText(value);
-            }
-        });
-    }
-    void bindToUserData(QComboBox *combo, QString key) {
-        connect(combo, static_cast<void (QComboBox::*)(int)>(&QComboBox::currentIndexChanged),
-                this, [this, combo, key](int index) {
-            this->event->put(key, combo->itemData(index).toString());
-        });
-    }
-
-signals:
-    void positionChanged(Event *event);
-    void xChanged(int);
-    void yChanged(int);
-    void elevationChanged(int);
-    void spriteChanged(QPixmap pixmap);
-    void onPropertyChanged(QString key, QString value);
-
-public slots:
-    void set_x(const QString &text) {
-        event->put("x", text);
-        updatePosition();
-    }
-    void set_y(const QString &text) {
-        event->put("y", text);
-        updatePosition();
-    }
-    void set_elevation(const QString &text) {
-        event->put("elevation", text);
-        updatePosition();
-    }
-    void set_sprite(const QString &text) {
-        event->put("sprite", text);
-        updatePixmap();
-    }
-    void set_script(const QString &text) {
-        event->put("script_label", text);
-    }
-
-protected:
-    void mousePressEvent(QGraphicsSceneMouseEvent*);
-    void mouseMoveEvent(QGraphicsSceneMouseEvent*);
-    void mouseReleaseEvent(QGraphicsSceneMouseEvent*);
-    void mouseDoubleClickEvent(QGraphicsSceneMouseEvent*);
 };
 
 #endif // EDITOR_H
